Manual merge of gotson:missing-year
Manually merge in the missing year functionality from gotson.

"This PR introduces a new command line argument -Y or --force-year:

when set, the transcode will happen like before, but the upload will be done by forcing the year of the torrent by using the one of the group. The problem with that is the new torrent will not be in the same group has the source of the transcode. This should require a manual action from staff to correct the newly added torrent and mark it as Unconfirmed.
when not set (default), the whole torrent will be skipped entirely, and won't be transcoded."